requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
pytest==8.4.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
